+++ library/riscv-aia/src/lib.rs
@@ -1,6 +1,9 @@
 #![doc = include_str!("../README.md")]
 #![no_std]
 
+#[macro_use]
+mod macros;
+
 mod iid;
 
 pub mod csrind;